+++ src/CG.Business/Strategies/ApplicationBuilderExtensions.cs
@@ -2,6 +2,7 @@
 using CG.Business.Properties;
 using CG.Configuration;
 using CG.Reflection;
+using CG.Runtime;
 using CG.Validations;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
@@ -137,12 +138,14 @@ namespace Microsoft.AspNetCore.Builder
             {
                 // If we get here then the configuration holds a single value.
 
+                var strategyName = configuration["Selected"];
+
                 try
                 {
                     // Process the selection.
                     applicationBuilder.ProcessStrategySection(
                         hostEnvironment,
-                        configuration["Selected"],
+                        strategyName,
                         configuration,
                         assemblyWhiteList,
                         assemblyBlackList
@@ -155,7 +158,7 @@ namespace Microsoft.AspNetCore.Builder
                         message: string.Format(
                             Resources.SingleFail,
                             nameof(UseStrategies),
-                            configuration["Selected"],
+                            strategyName,
                             configuration.GetPath()
                             ),
                         innerException: ex
@@ -201,15 +204,6 @@ namespace Microsoft.AspNetCore.Builder
            string assemblyBlackList
            )
         {
-            // Validate the parameters before attempting to use them.
-            Guard.Instance().ThrowIfNull(applicationBuilder, nameof(applicationBuilder))
-                .ThrowIfNull(configuration, nameof(configuration));
-
-            // Strategy name is a configuration parameter, not just a programming
-            //   argument to a method call, so, we'll check it here because the programmer
-            //   might have supplied the argument while the configuration itself might be 
-            //   missing any actual information.
-
             // Trim the strategy name, just in case.
             strategyName = strategyName.Trim();
 
@@ -249,13 +243,23 @@ namespace Microsoft.AspNetCore.Builder
                         //   loaded assembly, in case the caller fat-fingered the filename
                         //   in any way (the upcoming compares are all case sensitive).
 
+                        // Using a loader gives us a better ability to find assemblies
+                        //   at runtime - including dependent assemblies, which can be
+                        //   tricky, otherwise.
+                        var loader = new AssemblyLoader();
+
                         try
                         {
                             // Should we load the assembly by path, or name?
-                            if (assemblyNameOrPath.EndsWith(".dll"))
+                            if (assemblyNameOrPath.EndsWith(".dll", StringComparison.InvariantCultureIgnoreCase))
                             {
-                                // Load the assembly.
-                                var tempAsm = Assembly.LoadFrom(assemblyNameOrPath);
+                                // We'll need a complete path here.
+                                var completePath = Path.GetFullPath(
+                                    assemblyNameOrPath
+                                    );
+
+                                // Load the assembly by path.
+                                var tempAsm = loader.LoadFromAssemblyPath(completePath);
 
                                 // Add it to the white list.
                                 assemblyWhiteList = assemblyWhiteList.Length > 0
@@ -264,8 +268,10 @@ namespace Microsoft.AspNetCore.Builder
                             }
                             else
                             {
-                                // Load the assembly by yname.
-                                var tempAsm = Assembly.Load(assemblyNameOrPath);
+                                // Load the assembly by name.
+                                var tempAsm = loader.LoadFromAssemblyName(
+                                    new AssemblyName(assemblyNameOrPath)
+                                    );
 
                                 // Add it to the white list.
                                 assemblyWhiteList = assemblyWhiteList.Length > 0
@@ -291,7 +297,8 @@ namespace Microsoft.AspNetCore.Builder
             {
                 // If we get here then the 'Selected' field contains a value that
                 //   doesn't correspond to a child section. In itself, that isn't an
-                //   error, but, it probably is something we should warn about ...
+                //   error, since the child section it optional, but, it probably is
+                //   something we should warn about ...
                 // Hmm ... should we pull in an ILogger parameter just for this warning?
             }